+++ CHANGELOG.md
@@ -30,7 +30,7 @@ All notable changes to this project will be documented in this file.
     * [Full changelog: 1.3.0...1.4.0](https://github.com/ni/nidaqmx-python/compare/1.3.0...1.4.0)
 
 * ### Resolved Issues
-    * ...
+    * [820: Create a synchronization example for TDMS logging](https://github.com/ni/nidaqmx-python/issues/820)
 
 * ### Major Changes
     * Removed support for Python 3.9.